Homepage: Alerts card layout rework
